finishes planner rewrite with performance and legibility improvments
